requirements.txt
Flask>=2.0.0
Pillow>=9.0.0
reportlab>=3.6.0
docx2pdf>=0.1.8
openpyxl>=3.0.9
python-docx>=0.8.11